Router designed to download all user reviews
In router, a query is made to the database, where all user comments sorted by time are loaded. Next, you are requested to mark comments that the user liked and did not like. After that, the loaded notes are correlated with the previously received comments. The result should be a list of comments that can be edited immediately.