Filling with missing data
A cycle has been implemented that traverses the list of models and creates the necessary structure of properties using a previously generated template. If the product does not have the specified section, then it is created. Further, if there is no specified property, then the drain of 'emptiness' is put in its place.