Router for handling fetch requests
A file /crs/userRequest.js has been created in which the Router is connected to handle various user requests.
Router itself is included in the main /index.js file.
The crs / userRequest.js file has been created in which all fetch calls to the server will be registered. The file itself is included in the /views/partials/footer.hbs template.