Developed router catalog page
In router, two arrays are created in the session, where the id of the properties that the user has selected and the id of the properties that the user cannot select will be written. In this case, the prohibited properties belong to the same section as the properties selected by the user. The router also calls the filterProduct () method which will return a list of models with their fields and a filter list with parameters that will store the state of the property.