Request to get a list of properties that belong to the selected models
If the search results for the models did not give any result, then the query for allowed properties must be equal to the list of selected properties.
If the list of possible properties from the sorted models can be obtained, then this is done through a query, where all properties from the filter properties are obtained, but only those that belong to the sorted models.