Converts unlawful properties to selectable properties
If the section has a selected property, then it should be possible to select other properties from that section. However, if at the moment of selecting a property, next to it there are properties that cannot be selected, then they are added to the list of prohibitedOption (not allowed for selection). Therefore, if a property is in the list, then it should remain unavailable for selection. Also, if after updating the filter the property has received the status of allowed for selection, then it should be removed from the prohibitedOption list.