List of viewed products
Added the viewedProducts array to the "User" table, where the _id of the products that the user has viewed will be added.
In the product router, the product can be added to the list if the user is authorized. Also, in addition to adding a product to the list, the product's rating is increased by one.