The interface between menu and catalog has been implemented
A condition has been added to the interface, if router receives a link route with parameters, then it performs the following actions. Session filters are cleared. Next, a new filter is formed based on variables from parameters. After the formed filter, the route is redirected to the same router, now without parameters. This code on the catalog page will be executed once.