Mobile version of the menu
The file menuMobile.js has been created, which will contain the logic of the menu for the mobile version of the site.
Abolish the following link for backgroundCategory objects from the listType block on a touchend event for a touch screen. Cancellation of following a link is implemented using the preventDefault() method, which cancels all further events.