Initial conditions of the interface
Created a basic fulfillment condition only for Android and iPhone devices. On the mobile version, the menu interface starts with a parent category field, listCategory, which will fill the entire width of the screen. And the listType and listSearch blocks at the start position will be equal to zero.
Also, all roundingArrow elements must be hidden in the listCategory block.
Created a touchend event listener for backgroundCategory blocks from the global menu field.