Implemented sending a letter to change the password
The resetPassword.js file has been created where the template of the email sent to the user in case of a password change request is implemented.
In the file authorization.js, a call to the function of sending a letter is connected and a router for processing mail is added.
The method creates a token for the link and a date with the time until which the token will be valid. After that, all data is saved to the database.
At the end, the resetPas() method is called, which starts sending an email to the user with a link to reset the password.