Router password change
The authorization.js file implements a password change router. The method first checks if there is a user with a token that is not deprecated yet. If there is such a user, then it is checked for similarity to old passwords. If everything is successful, then the entered password is written to the database. Token and date are cleared. Everything is preserved.